+++ cmake/BLTMacros.cmake
@@ -482,11 +482,14 @@ macro(blt_patch_target)
     # PGI does not support -isystem
     if( (${arg_TREAT_INCLUDES_AS_SYSTEM}) AND (NOT "${CMAKE_CXX_COMPILER_ID}" STREQUAL "PGI"))
         get_target_property(_target_includes ${arg_NAME} INTERFACE_INCLUDE_DIRECTORIES)
-        if(_standard_lib_interface)
-            target_include_directories(${arg_NAME} SYSTEM ${_scope} ${_target_includes})
-        else()
-            set_property(TARGET ${arg_NAME} PROPERTY
-                         INTERFACE_SYSTEM_INCLUDE_DIRECTORIES ${_target_includes})
+        # Don't copy if the target had no include directories
+        if(_target_includes)
+            if(_standard_lib_interface)
+                target_include_directories(${arg_NAME} SYSTEM ${_scope} ${_target_includes})
+            else()
+                set_property(TARGET ${arg_NAME} PROPERTY
+                            INTERFACE_SYSTEM_INCLUDE_DIRECTORIES ${_target_includes})
+            endif()
         endif()
     endif()
 
@@ -525,12 +528,13 @@ endmacro(blt_patch_target)
 ##                     COMPILE_FLAGS [ flag1 [ flag2 ..]]
 ##                     LINK_FLAGS [ flag1 [ flag2 ..]]
 ##                     DEFINES [def1 [def2 ...]] 
-##                     GLOBAL [ON|OFF])
+##                     GLOBAL [ON|OFF]
+##                     EXPORTABLE [ON|OFF])
 ##
 ## Imports a library as a CMake target
 ##------------------------------------------------------------------------------
 macro(blt_import_library)
-    set(singleValueArgs NAME TREAT_INCLUDES_AS_SYSTEM GLOBAL)
+    set(singleValueArgs NAME TREAT_INCLUDES_AS_SYSTEM GLOBAL EXPORTABLE)
     set(multiValueArgs LIBRARIES
                        INCLUDES 
                        DEPENDS_ON
@@ -548,8 +552,12 @@ macro(blt_import_library)
         message(FATAL_ERROR "blt_import_library() must be called with argument NAME <name>")
     endif()
 
-    # Add all imported targets to a single imported interface target
-    if(${arg_GLOBAL})
+    if(${arg_EXPORTABLE})
+        if(${arg_GLOBAL})
+            message(FATAL_ERROR "blt_import_library: EXPORTABLE targets cannot be GLOBAL")
+        endif()
+        add_library(${arg_NAME} INTERFACE)
+    elseif(${arg_GLOBAL})
         add_library(${arg_NAME} INTERFACE IMPORTED GLOBAL)
     else()
         add_library(${arg_NAME} INTERFACE IMPORTED)