+++ cmake/thirdparty/SetupHIP.cmake
@@ -14,8 +14,6 @@ find_package(HIP REQUIRED)
 
 message(STATUS "HIP version:      ${HIP_VERSION_STRING}")
 message(STATUS "HIP platform:     ${HIP_PLATFORM}")
-#message(STATUS "HIP Include Path: ${HIP_INCLUDE_DIRS}")
-#message(STATUS "HIP Libraries:    ${HIP_LIBRARIES}")
 
 if(${HIP_PLATFORM} STREQUAL "hcc")
 	set(HIP_RUNTIME_DEFINE "__HIP_PLATFORM_HCC__")
@@ -34,10 +32,9 @@ set(HIP_RUNTIME_COMPILE_FLAGS "${HIP_RUNTIME_COMPILE_FLAGS};-D${HIP_RUNTIME_DEFI
 # depend on 'hip', if you need to use hip
 # headers, link to hip libs, and need to run your source
 # through a hip compiler (hipcc)
-blt_import_library(NAME      hip
-                   INCLUDES  ${HIP_INCLUDE_DIRS}
-                   LIBRARIES ${HIP_LIBRARIES}
-                   TREAT_INCLUDES_AS_SYSTEM ON)
+# This is currently used only as an indicator for blt_add_hip* -- FindHIP/hipcc will handle resolution
+# of all required HIP-related includes/libraries/flags.
+blt_import_library(NAME      hip)
 
 # depend on 'hip_runtime', if you only need to use hip
 # headers or link to hip libs, but don't need to run your source
@@ -47,4 +44,5 @@ blt_import_library(NAME          hip_runtime
                    DEFINES       ${HIP_RUNTIME_DEFINES}
                    COMPILE_FLAGS ${HIP_RUNTIME_COMPILE_FLAGS}
                    LIBRARIES     ${HIP_RUNTIME_LIBRARIES}
-                   TREAT_INCLUDES_AS_SYSTEM ON)
+                   TREAT_INCLUDES_AS_SYSTEM ON
+                   EXPORTABLE    ${BLT_EXPORT_THIRDPARTY})