+++ cmake/thirdparty/SetupMPI.cmake
@@ -182,4 +182,5 @@ blt_import_library(NAME          mpi
                    TREAT_INCLUDES_AS_SYSTEM ON
                    LIBRARIES     ${_mpi_libraries}
                    COMPILE_FLAGS ${_mpi_compile_flags}
-                   LINK_FLAGS    ${_mpi_link_flags} )
+                   LINK_FLAGS    ${_mpi_link_flags} 
+                   EXPORTABLE    ${BLT_EXPORT_THIRDPARTY})