+++ cmake/thirdparty/SetupThirdParty.cmake
@@ -98,6 +98,10 @@ blt_find_executable(NAME        Uncrustify
 blt_find_executable(NAME        Yapf
                     EXECUTABLES yapf)
 
+blt_find_executable(NAME        CMakeFormat
+                    EXECUTABLES cmake-format)
+
+
 #------------------------------------
 # Static analysis via Cppcheck
 #------------------------------------
@@ -114,4 +118,23 @@ if(CMAKE_GENERATOR STREQUAL "Unix Makefiles" OR CMAKE_GENERATOR STREQUAL "Ninja"
 
     blt_find_executable(NAME        ClangTidy
                         EXECUTABLES clang-tidy)
+endif()
+
+#------------------------------------
+# Code coverage
+#------------------------------------
+if (ENABLE_COVERAGE)
+    # Attempt to find the executables associated with gcov, lcov and genhtml.
+    # This requires that the associated features are enabled.
+    set(ENABLE_GCOV ON CACHE BOOL "")
+    set(ENABLE_LCOV ON CACHE BOOL "")
+    set(ENABLE_GENHTML ON CACHE BOOL "")
+    blt_find_executable(NAME        gcov
+                        EXECUTABLES gcov)
+
+    blt_find_executable(NAME        lcov
+                        EXECUTABLES lcov)
+
+    blt_find_executable(NAME        genhtml
+                        EXECUTABLES genhtml)
 endif()